refactor: テーマを切り替えつつスクリーンショットを撮る関数を作成
