+++ apps/site/src/__tests__/e2e/activities.test.ts
@@ -1,14 +1,16 @@
 import type { Page } from '@playwright/test'
 import { expect, test } from '@playwright/test'
 
+import { captureScreenshotWithTheme } from './utils/captureScreenshotWithTheme'
+
 const setup = async (page: Page) => {
   await page.goto('/activities')
 }
 
-test('/activities', async ({ page }) => {
+test('/activities', async ({ page, isMobile }) => {
   await setup(page)
 
-  await expect(page).toHaveScreenshot()
+  await captureScreenshotWithTheme(page, isMobile)
 })
 
 test('/activities 外部リンクは別タブで開く', async ({ page }) => {
