+++ apps/site/src/__tests__/e2e/index.test.ts-snapshots/-1-firefox-linux.png
