+++ apps/site/src/__tests__/e2e/index.test.ts-snapshots/-2-chromium-linux.png
